Add customisable `sinon` library
See https://github.com/lennym/reqres/issues/5

There is a need for users to be able to use a particular version of sinon for compatibility with other tools in their testing toolchain.

As such, expose a `reqres.sinon` property which can be used to set the `sinon` used internally by reqres to any user-provided library that implements `stub` and `spy`.